requirements.txt
discord.py 
blockcypher 
requests
